Move getAllTransitivelyReferencedResourceURNs to dependsOn
We want to use this function in invokes, move it to the dependsOn
package in preparation of that.
